#100: added toolbar colors framework
