+++ lib/ui/common/rich_text/rich_text.dart
@@ -1,2 +1,3 @@
 export './noted_rich_text_controller.dart';
 export './noted_rich_text_editor.dart';
+export './toolbar/noted_rich_text_toolbar.dart';
