+++ lib/ui/common/rich_text/toolbar/colors/noted_rich_text_toolbar_colors.dart
@@ -1,0 +1,13 @@
+part of '../noted_rich_text_toolbar.dart';
+
+class _ToolbarColorPicker extends StatefulWidget {
+  final NotedRichTextController controller;
+
+  const _ToolbarColorPicker({required this.controller, super.key});
+
+  @override
+  State<StatefulWidget> createState() {
+    // TODO: implement createState
+    throw UnimplementedError();
+  }
+}
