+++ lib/ui/common/rich_text/toolbar/noted_rich_text_toolbar.dart
@@ -1,17 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:noted_app/ui/common/rich_text/noted_rich_text_attributes.dart';
 import 'package:noted_app/ui/common/rich_text/noted_rich_text_controller.dart';
-import 'package:noted_app/ui/common/rich_text/toolbar/noted_rich_text_color_button.dart';
-import 'package:noted_app/ui/common/rich_text/toolbar/noted_rich_text_link_button.dart';
-import 'package:noted_app/ui/common/rich_text/toolbar/noted_rich_text_toggle_button.dart';
+import 'package:noted_app/ui/common/rich_text/toolbar/home/noted_rich_text_color_button.dart';
+import 'package:noted_app/ui/common/rich_text/toolbar/home/noted_rich_text_link_button.dart';
+import 'package:noted_app/ui/common/rich_text/toolbar/home/noted_rich_text_toggle_button.dart';
+
+part 'home/noted_rich_text_toolbar_home.dart';
+part 'colors/noted_rich_text_toolbar_colors.dart';
 
 enum _ToolbarState {
   home,
-  colorPicker,
+  colors,
 }
 
 Key _homeKey = Key(_ToolbarState.home.name);
-Key _colorPickerKey = Key(_ToolbarState.colorPicker.name);
+Key _colorsKey = Key(_ToolbarState.colors.name);
 
 class NotedRichTextToolbar extends StatefulWidget {
   final NotedRichTextController controller;
@@ -55,104 +58,9 @@ class _NotedRichTextToolbarState extends State<NotedRichTextToolbar> {
         ),
         child: switch (state) {
           _ToolbarState.home => _ToolbarHome(controller: widget.controller, key: _homeKey),
-          _ToolbarState.colorPicker => _ToolbarColorPicker(controller: widget.controller, key: _colorPickerKey),
+          _ToolbarState.colors => _ToolbarColorPicker(controller: widget.controller, key: _colorsKey),
         },
       ),
     );
-  }
-}
-
-class _ToolbarHome extends StatelessWidget {
-  final NotedRichTextController controller;
-
-  const _ToolbarHome({required this.controller, super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    ColorScheme colors = Theme.of(context).colorScheme;
-
-    return Wrap(
-      alignment: WrapAlignment.spaceBetween,
-      spacing: 4,
-      runSpacing: 6,
-      children: [
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.bold,
-          colors: colors,
-        ),
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.italic,
-          colors: colors,
-        ),
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.underline,
-          colors: colors,
-        ),
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.strikethrough,
-          colors: colors,
-        ),
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.h1,
-          colors: colors,
-        ),
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.h2,
-          colors: colors,
-        ),
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.h3,
-          colors: colors,
-        ),
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.ul,
-          colors: colors,
-        ),
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.ol,
-          colors: colors,
-        ),
-        NotedRichTextToggleButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.taskList,
-          colors: colors,
-        ),
-        NotedRichTextColorButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.textColor,
-          colors: colors,
-        ),
-        NotedRichTextColorButton(
-          controller: controller,
-          attribute: NotedRichTextAttribute.textBackground,
-          colors: colors,
-        ),
-        NotedRichTextLinkButton(
-          controller: controller,
-          colors: colors,
-        ),
-      ],
-    );
-  }
-}
-
-class _ToolbarColorPicker extends StatefulWidget {
-  final NotedRichTextController controller;
-
-  const _ToolbarColorPicker({required this.controller, super.key});
-
-  @override
-  State<StatefulWidget> createState() {
-    // TODO: implement createState
-    throw UnimplementedError();
   }
 }
